Upgrade to mlx-vlm 0.1.21
